pyproject.toml
[project]
name = "bothackathon"
version = "0.1.0"
requires-python = "~=3.14.0"
dependencies = [
    "pytelegrambotapi>=4.31.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "ipython>=9.10.0",
    "ruff>=0.15.2",
]
